fix memory leak in entity factory (#160)

+++ Editor/UI.cpp
@@ -320,7 +320,7 @@ namespace ToolKit
           0.23529413f, 0.24705884f, 0.25490198f, 0.00f};
       style->Colors[ImGuiCol_PopupBg] = {0.13f, 0.13f, 0.13f, 0.94f};
       style->Colors[ImGuiCol_Border]  = {
-           0.33333334f, 0.33333334f, 0.33333334f, 0.50f};
+          0.33333334f, 0.33333334f, 0.33333334f, 0.50f};
       style->Colors[ImGuiCol_BorderShadow] = {
           0.15686275f, 0.15686275f, 0.15686275f, 0.00f};
       style->Colors[ImGuiCol_FrameBg] = {
@@ -333,7 +333,7 @@ namespace ToolKit
       style->Colors[ImGuiCol_TitleBgCollapsed] = {0.16f, 0.29f, 0.48f, 1.00f};
       style->Colors[ImGuiCol_TitleBgActive]    = {0.00f, 0.00f, 0.00f, 0.51f};
       style->Colors[ImGuiCol_MenuBarBg]        = {
-                 0.27058825f, 0.28627452f, 0.2901961f, 0.80f};
+          0.27058825f, 0.28627452f, 0.2901961f, 0.80f};
       style->Colors[ImGuiCol_ScrollbarBg] = {
           0.27058825f, 0.28627452f, 0.2901961f, 0.60f};
       style->Colors[ImGuiCol_ScrollbarGrab] = {
@@ -347,7 +347,7 @@ namespace ToolKit
       style->Colors[ImGuiCol_SliderGrab]       = {0.70f, 0.70f, 0.70f, 0.62f};
       style->Colors[ImGuiCol_SliderGrabActive] = {0.30f, 0.30f, 0.30f, 0.84f};
       style->Colors[ImGuiCol_Button]           = {
-                    0.33333334f, 0.3529412f, 0.36078432f, 0.49f};
+          0.33333334f, 0.3529412f, 0.36078432f, 0.49f};
       style->Colors[ImGuiCol_ButtonHovered] = {
           0.21960786f, 0.30980393f, 0.41960788f, 1.00f};
       style->Colors[ImGuiCol_ButtonActive] = {
@@ -1399,7 +1399,7 @@ namespace ToolKit
           ImGui::Text("%s", m_msg.c_str());
         }
 
-        uint columnCount = m_buttons.size() + 2;
+        int columnCount = (int) m_buttons.size() + 2;
         columnCount += m_showCancel ? 1 : 0;
         // Center buttons.
         ImGui::BeginTable(
